(partial) added declare function
