+++ inc/tree.hpp
@@ -6,6 +6,8 @@
 #include <compare>
 #include <filesystem>
 #include <nlohmann/json.hpp>
+#include "crypt.hpp"
+#include "strops.hpp"
 
 using json = nlohmann::json;
 
@@ -46,6 +48,40 @@ block json_to_block(json input);
 block construct_block(std::string cont, std::unordered_set<std::string> p_hashes, int pow, std::string s_trip, unsigned long long set_time = get_raw_time(), std::string c_trip = "");
 
 std::vector<std::string> order_hashes(std::unordered_set<std::string> input_hashes);
+
+struct keypair {
+    std::string DSA;
+    std::string RSA;
+
+    keypair() {}
+    keypair(std::string dsak, std::string rsak) : DSA(dsak), RSA(rsak) {}
+};
+
+struct user {
+    std::string trip;
+    keypair pubkeys, prikeys;
+    // bool empty = false;
+
+    user() {
+        std::array<std::string, 2> DSA = cDSA::keygen();
+        std::array<std::string, 2> RSA = cRSA::keygen();
+
+        pubkeys = keypair(b64::encode(DSA[1]), b64::encode(RSA[1]));
+        prikeys = keypair(b64::encode(DSA[0]), b64::encode(RSA[0]));
+
+        this->trip = gen::trip(pubkeys.DSA + pubkeys.RSA, 24);
+    }
+
+    user(keypair pubset) : pubkeys(pubset) {
+        this->trip = gen::trip(pubset.DSA + pubset.RSA, 24);
+    }
+
+    user(keypair pubset, keypair priset) : pubkeys(pubset), prikeys(priset) {
+        this->trip = gen::trip(pubset.DSA + pubset.RSA, 24);
+    }
+
+    user(std::string trip, keypair priset) : trip(trip), prikeys(priset) {}
+};
 
 class Tree {
     private:
@@ -101,6 +137,9 @@ class Tree {
         std::unordered_set<std::string> get_parent_hash_union(std::unordered_set<std::string> c_hashes);
 
         std::vector<block> search_user(std::string trip = "");
+        bool declare_user(user user_, std::string nick = "");
+
+
         //std::vector<json> search(char message_type, std::string target_trip, std::string key, boost::function<bool(json)> filter = no_filter, int start_b = -1, int end_b = -1);
 
         bool verify_chain();
